+++ .env.example
@@ -1,0 +1,5 @@
+# Exactly one of the following variables must be set:
+# YAML_FILEPATH
+# POSTGRES_CONNECTION_STRING
+# An error will be thrown if neither or both are set.
+YAML_FILEPATH=gateway-endpoints.yaml # The local path to the YAML file containing the GatewayEndpoints.
