+++ main.go
@@ -15,6 +15,8 @@ import (
 
 	grpc_server "github.com/buildwithgrove/path-auth-data-server/grpc"
 	"github.com/buildwithgrove/path-auth-data-server/yaml"
+
+	_ "github.com/joho/godotenv/autoload"
 )
 
 const (
@@ -63,7 +65,7 @@ func main() {
 
 	// Create a new HTTP handler that serves both gRPC and HTTP
 	grpcAndHTTPHandler := h2c.NewHandler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.ProtoMajor == 2 && r.Header.Get("Content-Type") == "application/grpc" {
+		if isRequestGRPC(r) {
 			grpcServer.ServeHTTP(w, r)
 		} else {
 			mux.ServeHTTP(w, r)
@@ -77,4 +79,9 @@ func main() {
 	if err := httpServer.Serve(lis); err != nil {
 		panic(fmt.Sprintf("failed to serve: %v", err))
 	}
+}
+
+// isRequestGRPC checks the true if the request is a gRPC request by checking the protocol and content type.
+func isRequestGRPC(req *http.Request) bool {
+	return req.ProtoMajor == 2 && req.Header.Get("Content-Type") == "application/grpc"
 }
